--- a/src/lib.rs
+++ b/src/lib.rs
@@ -21,6 +21,7 @@ mod treefrog;
 pub use treefrog::{
     extend_anti::ExtendAnti, extend_with::ExtendWith, filter_anti::FilterAnti,
     filter_with::FilterWith, Leaper, RelationLeaper,
+    filters::{PrefixFilter, ValueFilter},
 };
 
 /// A static, ordered list of key-value pairs.