This PR introduces `treefrog::filters::{PrefixFilter, ValueFilter}`, each of which implement `Leaper`. The first is constructed from a `fn(&Tuple) -> bool` and the second from a `fn(&Tuple, &Val) -> bool`. Each of them do the very simple action you would expect, the first one ruling out extensions based on `prefix`  and the second filtering `values` down using the predicate.