@nikomatsakis 2018-12-05
Seems like a handi concept, though I'm not sure that I have a use for it yet =)